changed stuff

+++ src/interrupts/constants.rs
@@ -1,4 +1,4 @@
 extern "C" {
-    pub static SOFTWARE_INTERRUPT: usize;
-    pub static SOFTWARE_VAL_ENABLE_MIE: usize;
+    pub static SOFTWARE_INTERRUPT: u32;
+    pub static SOFTWARE_VAL_ENABLE_MIE: u32;
 }
